virtual_scientist_system_langgraph/requirements.txt
Flask
requests
langchain
langchain-groq
langgraph
psycopg2-binary
python-dotenv